Migrate events_subscribe tests
